requirements.txt
azure-functions
azure-ai-formrecognizer>=3.3.0
google-generativeai>=0.8.3
requests>=2.31.0
openai>=1.30.0
azure-storage-blob>=12.19.0

